String Data Type.py
String data types represent and manipulate text-based data in programming. They are used for storing and working with sequences of characters. Strings offer operations like concatenation and comparison. They are widely used for tasks such as user input and text processing. Strings are usually immutable, and encoding and length considerations are important.